requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-markdown-deux==1.0.5
django-markdown2==0.3.1
django-pagedown==2.1.3
django-videofield==0.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
markdown2==2.3.10
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
